(finally) have instruction cache tag load op finished

+++ src/src/n64/icache_tests.rs
@@ -7,8 +7,8 @@ mod icache_tests
     fn can_translate_virtual_address_to_cache_line() {
         let mut icache: ICache = ICache::new();
 
-        let line_index_high: u32 = icache.parse_line_index_from_virtual_index(0x00003FE0);
-        let line_index_low: u32 = icache.parse_line_index_from_virtual_index(0xFFFFC01F);
+        let line_index_high: usize = icache.parse_line_index_from_virtual_index(0x00003FE0);
+        let line_index_low: usize = icache.parse_line_index_from_virtual_index(0xFFFFC01F);
 
 
         assert_eq!(511, line_index_high);
